backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
supabase==2.7.4
openai==1.45.0
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
langchain-text-splitters==0.3.0
python-jose==3.3.0
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
